Remove abstract class exports from API, and rename 'assignment.ts' module to 'assignment_expression.ts'.
